Compiles and runs (I think). See below.
Current state: I have added the appropriate methods to FilletBase to add the correct data
to the PropertyFilletEdges thing. I've also extended TopoNamingHelper and TopoShape such
that FilletBase has sufficient access to the TNaming stuff to store enough info, in the
form of a string. What still needs to be done is to add sufficient python functionality to
FilleBase such that the Mod/Part/Gui/DlgFilletEdges thing can call 'Fillet.setEdge' or
'Fillet.setEdges' appropriately.