Fixed selectEdges. See below for TODO
Now, the only thing that is left is to actually use the edgetag that is now stored in the
FilletElement to retrieve the appropriate edge, rather than using the edgeid. I'm leaving
the edgeid in there for now, just in case there's code I don't know about that uses it.