IT WORKS!!! See below for next steps.
So, in TopoNamingHelper::TrackModifiedShape, I need to find a way to actually add only the
modified Faces as opposed to hard-coding it like I have right now. Also, maybe other
stuff. Look hard for "TODO" in the code. Next step is to see how we're doing with a Fused
shape...